feat: profile page with user info  (#28)
* feat: reduced shadow effect to md

* feat: logo should be clickable and link to dashboard

* feat: sign out button logs out user

* feat: make the navbar fixed to the top of the screen

Fixes an issue that the viewport is extending the page even when there is enough content.

* fix: add padding to avoid the navbar from covering component

* fix: add padding to prevent the navbar from covering the component

* chore: moved register view to separate file

* feat: serializing User for access to profile data

Temporarily use all fields to get a full object and test what all the fields are and what we can use them for.

* chore: Moved the login view to a separate file

* feat: creating user view to get profile information

* chore: remove views file and user views directory

* feat: update icon for journal title edit

* feat: profile page initial setup

* fix: change "currentUser" to "me" in api url

* fix: missing trailing slash on me route

* fix: is_valid() not meant for get

* fix: specify desired fields for User

* chore: remove redundant react import

* refactor: renamed to authService

* feat: service method to get user profile

* feat: general setup of the profile section done

The inputs are set to show up if the flag is true for editing. Though since not all the fields are meant to be updated, I think a modal is better for this. It looks tacky to only have some change to input.

* feat: profile modal to update profile

* fix: updated name for profile url

* feat: PUT method for updating profile

* fix: prevent username from being updated

* feat: add username as part of modal, but disable access

Username is not mutable, but it should be there to keep the modal symmetric and is clearly darker to show that it can't be changed.

* feat: update user profile function

* feat: closable pop up when there is an error