May 26: Home Dashboard (#31)
* fix: use __all__ instead of listing all attributes

* chore: remove action on forms

The "action='#'" was part of the tailwind html code that I built off of.

* feat: service class for notes

* feat: note modal to get input for a new note

* feat: displaying the modal on click

* fix: precise error messages

* fix: throw error on login

Login errors were being sent to console, not being thrown. The Login page expects a thrown error, which can be displayed to the user.

* fix: using default data to create a journal

If a user does not have a journal and a get request is made for it, it is made using default data. This journal can later be updated as the user wishes.

* fix: use const arrow function

* fix: change to showModal needs lambda

* feat: overhaul create modal

* fix: trim journal title before update

* fix: trim username and password before registering

* feat: modal customized & saves the data from input

The modal is now clear and shows purpose for our usecase. the data is cleaned and validated before being added as part of the note data. The data is sent to the save function passed through props, which saves the note and adds it to the list.

* feat: sends a request to create a note

* feat: service functions to update and get all notes

* fix: conditionally render optional fields

* fix: use existing note data to fill in note state

* fix: update props sent to NoteModal

Sending the note to be edited and a function to handle the deletion of a note. The state that tracks the selected note also must be set to false when the modal is closed or submitted.

* chore: moved frontend to port 3000

* fix: noteId should be passed in url

* feat: separate the save process depending on whether there was a selected note

Given a selected note, it should be updated. Otherwise, we are creating a new note.

* fix: lighter border color on the note

* feat: sort notes on update, create, and original fetch

Notes are sorted by when they were last updated, with most recently updated notes are higher.

* chore: remove unused import

* chore: move formatDate to a util file

* chore: remove the button to check the data for development

* fix: formatDate was not properly imported

* feat: added service function to delete a note

* feat: added delete button for deleting note, updated update button

Using @heroicon svg's for the create and delete icons. Delete button will submit the form if not explicitly set to type="button". Call the onDelete function when clicked.

* feat: add function to handle the delete flow

* feat: updating shadow and border styles

* feat: profile page with user info  (#28)

* feat: reduced shadow effect to md

* feat: logo should be clickable and link to dashboard

* feat: sign out button logs out user

* feat: make the navbar fixed to the top of the screen

Fixes an issue that the viewport is extending the page even when there is enough content.

* fix: add padding to avoid the navbar from covering component

* fix: add padding to prevent the navbar from covering the component

* chore: moved register view to separate file

* feat: serializing User for access to profile data

Temporarily use all fields to get a full object and test what all the fields are and what we can use them for.

* chore: Moved the login view to a separate file

* feat: creating user view to get profile information

* chore: remove views file and user views directory

* feat: update icon for journal title edit

* feat: profile page initial setup

* fix: change "currentUser" to "me" in api url

* fix: missing trailing slash on me route

* fix: is_valid() not meant for get

* fix: specify desired fields for User

* chore: remove redundant react import

* refactor: renamed to authService

* feat: service method to get user profile

* feat: general setup of the profile section done

The inputs are set to show up if the flag is true for editing. Though since not all the fields are meant to be updated, I think a modal is better for this. It looks tacky to only have some change to input.

* feat: profile modal to update profile

* fix: updated name for profile url

* feat: PUT method for updating profile

* fix: prevent username from being updated

* feat: add username as part of modal, but disable access

Username is not mutable, but it should be there to keep the modal symmetric and is clearly darker to show that it can't be changed.

* feat: update user profile function

* feat: closable pop up when there is an error

* feat: home dashboard with recent notes widget (#30)

* feat: trash icon on Note Modal instead of X

* feat: applied quicksand font

* feat: updating the update/create note icon

* fix: increased line-clamp to 4

* feat: up to four most recent notes displayed

The notes are clickable, but only send to the journal page. This is more or less fine since the notes are the top four notes. The chevron also redirects to the journal page.

* chore: comment out variable that is currently unused

* fix: not using journal object on home page

* chore: remove console logs